TF05/api/requirements.txt
Flask==3.0.0
PyYAML==6.0.1
requests==2.31.0
PyMySQL==1.1.0